refactor: Pageable 사용하도록 수정 (#141)

+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/post/presentation/dto/HomeFeedRequest.java
@@ -18,11 +18,11 @@ public class HomeFeedRequest {
         return appUser;
     }
 
-    public Long getPage() {
-        return page;
+    public int getPage() {
+        return Math.toIntExact(page);
     }
 
-    public Long getLimit() {
-        return limit;
+    public int getLimit() {
+        return Math.toIntExact(limit);
     }
 }
